+++ api/gin/headers.go
@@ -1,0 +1,10 @@
+package gin
+
+const (
+	HeaderAuthorization  = "Authorization"
+	HeaderContentType    = "Content-Type"
+	HeaderXSHA256        = "X-SHA-256"
+	HeaderXContentType   = "X-Content-Type"
+	HeaderXContentLength = "X-Content-Length"
+	HeaderXUploadMessage = "X-Upload-Message"
+)
